Paring down some things - bit more to go.
